+++ doc/release_notes.md
@@ -2,10 +2,13 @@
 ## Upcoming next release
 - Encrpyted SQL*Net connection if server has SQLNET.ENCRYPTION_SERVER != rejected<br>
   Requires release of v6.1.7 for https://github.com/rsim/oracle-enhanced as precondition (https://github.com/rsim/oracle-enhanced/pull/2284)
-- ROWID values encapsulated in ROWIDTOCHAR() function for before JSON_OBJECT
+
+## 2023-06-15 Release 1.10.16
+- API functions /db_triggers/generate and /db_triggers/generate_all return http status code 207 (Multi-Status) if DB errors occur at trigger generation
 
 ## 2023-06-15 Release 1.10.15
 - Bugfix: Ensure timestamp fraction delimiter for Oracle release < 19.1 is always a dot no matter which language setting is active
+- ROWID values encapsulated in ROWIDTOCHAR() function for before JSON_OBJECT
 
 ## 2022-12-13 Release 1.10.11
 - Bugfix: wrong URL for css file after docker start with suffixed URL